Extract common functionality for linux shared folder mount
Provide a base class for linux style shared folder mounting. This allows
reuse of the uid/gid discovery logic as well as the upstart event notification
