+++ plugins/guests/linux/plugin.rb
@@ -1,4 +1,5 @@
 require "vagrant"
+require_relative "cap/mount_shared_folder_base"
 
 module VagrantPlugins
   module GuestLinux
